poc finalizada

+++ app/Livewire/Usuario/ExcluirUsuario.php
@@ -21,6 +21,11 @@ class ExcluirUsuario extends Component
 
     public function excluir(): void
     {
+        if ($this->usuario->empresas()->count() > 0) {
+            $this->js("alert('Usuário com empresa não poderá ser excluído!')");
+            return;
+        }
+
         $this->usuario->delete();
 
         $this->js("alert('Usuário excluído com sucesso!')");
